Move agent3 to agent
