+++ agent/src/main/java/com/microsoft/applicationinsights/agent3/dev/DevLogger.java
