+++ agent/src/main/java/com/microsoft/applicationinsights/agent3/model/QuerySpanImpl.java
