+++ agent/src/main/java/com/microsoft/applicationinsights/agent3/utils/LoggerSpans.java
