+++ settings.gradle
@@ -20,8 +20,8 @@
  */
 
 
+include 'agent'
 include 'agent-old'
-include 'agent3'
 include 'core'
 include 'logging:log4j1_2'
 include 'logging:log4j2'
